Removed the chans package
