+++ batch.go
@@ -3,7 +3,7 @@ package rill
 import (
 	"time"
 
-	"github.com/destel/rill/chans"
+	"github.com/destel/rill/internal/common"
 )
 
 // Batch groups items from an input channel into batches based on a maximum size and a timeout.
@@ -12,13 +12,13 @@ import (
 // The timeout countdown starts when the first item is added to a new batch.
 func Batch[A any](in <-chan Try[A], n int, timeout time.Duration) <-chan Try[[]A] {
 	values, errs := Unwrap(in)
-	batches := chans.Batch(values, n, timeout)
+	batches := common.Batch(values, n, timeout)
 	return WrapAsync(batches, errs)
 }
 
 // Unbatch is the inverse of Batch. It takes a channel of batches and emits individual items.
 func Unbatch[A any](in <-chan Try[[]A]) <-chan Try[A] {
 	batches, errs := Unwrap(in)
-	values := chans.Unbatch(batches)
+	values := common.Unbatch(batches)
 	return WrapAsync(values, errs)
 }
