+++ internal/common/core_test.go
@@ -3,7 +3,6 @@ package common
 import (
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/destel/rill/internal/th"
 )
@@ -30,7 +29,7 @@ func TestMapOrFilter(t *testing.T) {
 					return fmt.Sprintf("%03d", x), x%2 == 0
 				})
 
-				outSlice := toSlice(out)
+				outSlice := th.ToSlice(out)
 
 				expectedSlice := make([]string, 0, 20)
 				for i := 0; i < 20; i++ {
@@ -51,91 +50,7 @@ func TestMapOrFilter(t *testing.T) {
 					return x, x%2 == 0
 				})
 
-				outSlice := toSlice(out)
-
-				if ord || n == 1 {
-					th.ExpectSorted(t, outSlice)
-				} else {
-					th.ExpectUnsorted(t, outSlice)
-				}
-			})
-
-		}
-	})
-}
-
-func universalMapOrFlatMap[A, B any](ord bool, in <-chan A, n int, f func(A) (b B, bb <-chan B, flat bool)) <-chan B {
-	if ord {
-		return OrderedMapOrFlatMap(in, n, f)
-	}
-	return MapOrFlatMap(in, n, f)
-}
-
-func TestMapOrFlatMap(t *testing.T) {
-	th.TestBothOrderings(t, func(t *testing.T, ord bool) {
-		for _, n := range []int{1, 5} {
-
-			t.Run(th.Name("nil", n), func(t *testing.T) {
-				out := universalMapOrFlatMap(ord, nil, n, func(x int) (int, <-chan int, bool) { return x, nil, true })
-				th.ExpectValue(t, out, nil)
-			})
-
-			t.Run(th.Name("correctness", n), func(t *testing.T) {
-				in := th.FromRange(0, 20)
-				out := universalMapOrFlatMap(ord, in, n, func(x int) (string, <-chan string, bool) {
-					if x%2 == 0 {
-						return fmt.Sprintf("%03dA", x), nil, false
-					} else {
-						res := fromSlice([]string{fmt.Sprintf("%03dB", x), fmt.Sprintf("%03dC", x)})
-						return "dummy", res, true
-					}
-				})
-
-				outSlice := toSlice(out)
-
-				expectedSlice := make([]string, 0, 2*20)
-				for i := 0; i < 20; i++ {
-					if i%2 == 0 {
-						expectedSlice = append(expectedSlice, fmt.Sprintf("%03dA", i))
-					} else {
-						expectedSlice = append(expectedSlice, fmt.Sprintf("%03dB", i), fmt.Sprintf("%03dC", i))
-					}
-				}
-
-				th.Sort(outSlice)
-				th.ExpectSlice(t, outSlice, expectedSlice)
-			})
-
-			t.Run(th.Name("nil hang", n), func(t *testing.T) {
-				in := th.FromRange(0, 20)
-				out := universalMapOrFlatMap(ord, in, n, func(x int) (string, <-chan string, bool) {
-					if x == 10 {
-						return "", nil, true
-					} else {
-						return "", fromSlice([]string{"B", "C"}), true
-					}
-				})
-
-				th.ExpectNeverClosedChan(t, out, 1*time.Second)
-			})
-
-			t.Run(th.Name("ordering", n), func(t *testing.T) {
-				in := th.FromRange(0, 20000)
-
-				out := universalMapOrFlatMap(ord, in, n, func(x int) (string, <-chan string, bool) {
-					if x%2 == 0 {
-						return fmt.Sprintf("%06dA", x), nil, false
-					} else {
-						res := fromSlice([]string{
-							fmt.Sprintf("%06dB", x),
-							fmt.Sprintf("%06dC", x),
-							fmt.Sprintf("%06dD", x),
-						})
-						return "", res, true
-					}
-				})
-
-				outSlice := toSlice(out)
+				outSlice := th.ToSlice(out)
 
 				if ord || n == 1 {
 					th.ExpectSorted(t, outSlice)
@@ -178,7 +93,7 @@ func TestOrderedMapAndSplit(t *testing.T) {
 
 					outSlices := make([][]string, numOuts)
 					th.DoConcurrentlyN(numOuts, func(i int) {
-						outSlices[i] = toSlice(outs[i])
+						outSlices[i] = th.ToSlice(outs[i])
 					})
 
 					expectedSlices := make([][]string, 3)
@@ -207,7 +122,7 @@ func TestOrderedMapAndSplit(t *testing.T) {
 
 					outSlices := make([][]string, numOuts)
 					th.DoConcurrentlyN(numOuts, func(i int) {
-						outSlices[i] = toSlice(outs[i])
+						outSlices[i] = th.ToSlice(outs[i])
 					})
 
 					for i := range outSlices {
