--- internal/common/helpers_test.go
@@ -1,18 +1,0 @@
-package common
-
-func fromSlice[A any](slice []A) <-chan A {
-	out := make(chan A, len(slice))
-	for _, a := range slice {
-		out <- a
-	}
-	close(out)
-	return out
-}
-
-func toSlice[A any](in <-chan A) []A {
-	var res []A
-	for x := range in {
-		res = append(res, x)
-	}
-	return res
-}
