+++ util.go
@@ -1,17 +1,15 @@
 package rill
 
-import (
-	"github.com/destel/rill/chans"
-)
+import "github.com/destel/rill/internal/common"
 
 // Drain consumes and discards all items from an input channel, blocking until the channel is closed
 func Drain[A any](in <-chan A) {
-	chans.Drain(in)
+	common.Drain(in)
 }
 
 // DrainNB is a non-blocking version of [Drain].
 func DrainNB[A any](in <-chan A) {
-	chans.DrainNB(in)
+	common.DrainNB(in)
 }
 
 // Buffer takes a channel of items and returns a buffered channel of exact same items in the same order.
@@ -22,7 +20,7 @@ func DrainNB[A any](in <-chan A) {
 //	ids = Buffer(ids, 100)
 //	// Now up to 100 ids can be buffered if subsequent stages of the pipeline are slow
 func Buffer[A any](in <-chan A, n int) <-chan A {
-	return chans.Buffer(in, n)
+	return common.Buffer(in, n)
 }
 
 // FromSlice converts a slice into a channel.
@@ -43,7 +41,7 @@ func ToSlice[A any](in <-chan Try[A]) ([]A, error) {
 
 	for x := range in {
 		if err := x.Error; err != nil {
-			chans.DrainNB(in)
+			DrainNB(in)
 			return res, err
 		}
 		res = append(res, x.V)
